Refactor jwt.js into different layers
