+++ user-service/controller/user-controller.js
@@ -1,7 +1,8 @@
 import 'dotenv/config';
+import jwt from 'jsonwebtoken';
 
 import { ormCreateUser as _createUser, authenticateUser, isExistingUser, isExistingEmail, ormUpdateAccount as _updateAccount, ormDeleteAccount } from '../model/user-orm.js';
-import { blacklistJwt, generateJwt, decodeJwt } from '../model/jwt.js';
+import { ormBlacklistJwt as _blacklistJwt, ormGenerateJwt as _generateJwt, ormDecodeJwt as _decodeJwt, ormGetBlacklistedJwt as _getBlacklistedJwt } from '../model/jwt-orm.js';
 
 export async function createUser(req, res) {
     try {
@@ -24,7 +25,7 @@ export async function createUser(req, res) {
                 return res.status(201).json({message: `Created new user ${username} successfully!`});
             }
         } else {
-            return res.status(400).json({message: 'Username and/or Password are missing!'});
+            return res.status(400).json({message: 'Username/Email/Password are missing!'});
         }
     } catch (err) {
         return res.status(500).json({message: 'Database failure when creating new user!'})
@@ -42,7 +43,7 @@ export async function signin(req, res) {
                 return res.status(401).json({ message: 'User does not exist and/or wrong password.' });
             }
             // Create JWT
-            const token = generateJwt(signedInUser);
+            const token = _generateJwt(signedInUser);
             // Send cookie
             res.cookie('token', token, { httpOnly: true });
             return res.status(200).json({ token });
@@ -56,7 +57,7 @@ export async function signin(req, res) {
 
 export async function logout(req, res) {
     const { token } = req;
-    await blacklistJwt(token);
+    await _blacklistJwt(token);
     res.clearCookie('token');
     return res.status(200).json({ message: 'Logout successful.'})
 }
@@ -64,7 +65,7 @@ export async function logout(req, res) {
 export async function updateAccount(req, res) {
     const { token } = req;
     const { newUsername, newPassword } = req.body;
-    const { username } = decodeJwt(token);
+    const { username } = _decodeJwt(token);
     const newProfile = {};
 
     // Only add to the new profile to update to if the value exists
@@ -94,7 +95,7 @@ export async function updateAccount(req, res) {
           return res.status(401).json({ message: 'User does not exist and/or wrong password.' });
       }
       // Create new JWT for updated profile
-      const token = generateJwt(updatedUser);
+      const token = _generateJwt(updatedUser);
       // Send cookie
       res.cookie('token', token, { httpOnly: true });
       return res.status(200).json({ message: 'Updated account successfully.' });
@@ -105,8 +106,8 @@ export async function updateAccount(req, res) {
 
 export async function deleteAccount(req, res) {
     const { token } = req;
-    const { username } = decodeJwt(token);
-    await blacklistJwt(token);
+    const { username } = _decodeJwt(token);
+    await _blacklistJwt(token);
     res.clearCookie('token');
     if (await ormDeleteAccount(username)) {
         return res.status(200).json({ message: `Deleted account ${username} successfully.` });
@@ -117,6 +118,27 @@ export async function deleteAccount(req, res) {
 
 export function acknowledgeJWTValidity(req, res) {
     const { token } = req;
-    const { username } = decodeJwt(token);
+    const { username } = _decodeJwt(token);
     return res.status(200).json({ username, message: 'Valid JWT!' });
+}
+
+export async function authenticateJwt(req, res, next) {
+    const token = req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ message: 'No token provided.' });
+    }
+
+    const isBlacklisted = await _getBlacklistedJwt(token);
+    if (isBlacklisted) {
+        return res.status(401).json({ message: 'Please sign in again.' });
+    }
+
+    jwt.verify(token, process.env.JWT_SECRET_KEY, (error, user) => {
+        if (error) {
+            return res.status(401).json({ message: 'Invalid token.' })
+        }
+
+        req.token = token;
+        next();
+    })
 }
